BSCar/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Werkzeug==3.0.4
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
pytest==8.3.3
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
